+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Solspace Express Forms Changelog
 
+## 1.0.0-beta.3 - 2018-04-23
+### Changes
+- Various improvements to developer events.
+
+### Fixed
+- Fixed a bug where CP could show errors if an incorrect template path was specified for Email Notification Templates.
+- Fixed a bug where installing Express Forms could error in some cases.
+
 ## 1.0.0-beta.2 - 2018-04-19
 ### Added
 - Added support for Opportunity resource mapping in Salesforce API integration (Pro).